refactor: pr review stuff
